+++ crates/revelio-parsers-audio/src/usac.rs
@@ -5,7 +5,7 @@ use revelio_core::{FileAnalyze, StreamKind};
 /// Detection: ADTS 0xFFF with extended AudioObjectType 42.
 /// Fills: Profile, sample rate, channels.
 pub fn parse_usac(fa: &mut FileAnalyze) -> bool {
-    let buf = match fa.peek_raw(fa.remain() as usize) {
+    let buf = match fa.peek_raw(fa.remain()) {
         Some(b) => b,
         None => return false,
     };
