+++ crates/revelio-parsers-image/src/heif.rs
@@ -4,7 +4,7 @@ use revelio_core::{FileAnalyze, StreamKind};
 /// HEIF files have major_brand "mif1", "msf1", "heic", "heix", "hevc",
 /// "heim", "heis", "hevm", "hevs".
 pub fn parse_heif(fa: &mut FileAnalyze) -> bool {
-    let buf = fa.peek_raw(fa.remain() as usize).map(|b| b.to_vec());
+    let buf = fa.peek_raw(fa.remain()).map(|b| b.to_vec());
     let Some(buf) = buf else { return false };
     if buf.len() < 12 { return false; }
