+++ crates/revelio-parsers-video/src/avs3.rs
@@ -4,7 +4,7 @@ use revelio_core::{FileAnalyze, StreamKind};
 /// Detection: Annex B 0x000001B0 NAL.
 /// Fills: Profile, dimensions, bit depth.
 pub fn parse_avs3(fa: &mut FileAnalyze) -> bool {
-    let buf = fa.peek_raw(fa.remain() as usize).map(|b| b.to_vec());
+    let buf = fa.peek_raw(fa.remain()).map(|b| b.to_vec());
     let Some(buf) = buf else { return false };
     if buf.len() < 4 { return false; }
     if &buf[0..4] == b"AVS3" || (buf.len() >= 5 && buf[0] == 0x00 && buf[1] == 0x00 && buf[2] == 0x00 && buf[3] == 0x01 && buf[4] == 0xB0) {
